requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
build==1.2.1
CacheControl==0.14.0
celery==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
colorlog==6.8.2
crashtest==0.4.1
cron-descriptor==1.4.3
distlib==0.3.8
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.2
django-ipware==7.0.1
django-redis==5.4.0
django-split-settings==1.3.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
gunicorn==22.0.0
identify==2.6.0
idna==3.7
importlib_metadata==8.0.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
kombu==5.3.7
more-itertools==10.3.0
msgpack==1.0.8
nodeenv==1.9.1
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
pre-commit==3.7.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PyJWT==2.8.0
pyproject_hooks==1.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ipware==3.0.0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.4
redis==5.0.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.0
supervisor==4.2.5
tomli==2.0.1
tomlkit==0.13.0
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
whitenoise==6.7.0
zipp==3.19.2
